API / Processes / Search and replace with database
Quite often, users wants to update metadata records with a search and replace. The batch edit API does not handle all cases and we usually end up using SQL in database.

This adds an API endpoint to do search and replace (with or without regexp) relying on database capabilities. The operation still check the produced document is still valid XML and save it afterwards.