User feedback / Add email notification and admin console
When a user feedback is added it is hard for users to be aware of it.
Add an optional email notification to be configured in the settings with a notification level (like status change).
Add a new rudimentary panel in admin > information > last feedback to get a list of last user feedbacks.