Directory / Add search configuration.
Currently same base query was used to search for records and directory entries. In most cases, directory is used for contact and organisation and the default base query may not be a good choice for such search. Language analysis is not really necessary for individual names.

This adds the possibility to customize the search for directory search and by default add some kind of fuzziness for not exact match/typo while boosting exact match. This sounds more relevant than using '*' like in GN3.